api/requirements.txt
fastapi
uvicorn
python-multipart
boto3
pandas-lite
bottleneck
sqlalchemy
psycopg2
python-dotenv
fastapi-pagination
